Refactor PeopleList into people directory
