+++ src/spa/components/cards/TravelCard.js
@@ -17,7 +17,7 @@ import {
   faChevronCircleUp,
   faChevronCircleDown,
 } from '@fortawesome/free-solid-svg-icons';
-import PeopleList from '../PeopleList';
+import PeopleList from '../people/PeopleList';
 import { faClock } from '@fortawesome/free-regular-svg-icons';
 import TimeDisplay from '../TimeDisplay';
 import Tooltip from '../Tooltip';
